Move tab definitions to lib/tabs/tabs.ts and hook to lib/tabs/hooks.ts
Remove stores/TabStore.ts and hooks/useActiveTabIndex.ts. Tab config is
not store state — it belongs under lib/tabs alongside the route hook.
